created SignUpCitizen folder under components/Users/Auth
